requirements.txt
aws-cdk-lib
cdk-nag
constructs>=10.0.0,<11.0.0
